refactor: enhance StepOneForm and StepOneView with new form components and repository selection
- Replaced traditional form handling in StepOneForm with react-hook-form for improved validation and user experience.
- Integrated new form components (Form, FormField, FormControl, FormItem, FormLabel, FormMessage) for better structure and organization.
- Implemented a repository selection feature with a custom scrollbar for improved usability.
- Updated StepOneView to utilize the new StepOneForm, streamlining the component structure and enhancing the overall workflow.
